// frontend/src/app/core/generated/circabc/model/archiveNode.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Attachment } from './attachment';


export interface ArchiveNode { 
    id?: string;
    type?: string;
    service?: ArchiveNode.ServiceEnum;
    name?: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title?: { [key: string]: string; };
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    description?: { [key: string]: string; };
    properties?: { [key: string]: string; };
    /**
     * Representation of the permissions given to the user making the call. To know what are the permissions for the specific node, the permission definition may vary depending on its type or service. 
     */
    permissions?: { [key: string]: string; };
    attachments?: Array<Attachment>;
    parentId?: string;
    notifications?: string;
    /**
     * true if the current user faved the node 
     */
    favourite?: boolean;
    hasSubFolders?: boolean;
    hasGuestAccess?: boolean;
    /**
     * For nodes created by the CIRCABC migration/import process, the NodeRef of the original (source) node they were migrated from (stored via the ci:migrated aspect). Absent for non-migrated nodes. 
     */
    originalNodeRef?: string;
    deletedBy?: string;
    deletedDate?: string;
}
export namespace ArchiveNode {
    export const ServiceEnum = {
        Library: 'library',
        Information: 'information',
        Events: 'events',
        Newsgroups: 'newsgroups',
        Directory: 'directory'
    } as const;
    export type ServiceEnum = typeof ServiceEnum[keyof typeof ServiceEnum];
}